Removed Assistant, updated coach description
We no longer use the term assistance, everyone supporting students during the workshop is a coach.  Having separation between teachers and assistants was artificial and very few people wanted the percieved responsibility of a teacher.  Having more of a one-2-one ratio between students and coaches helps with this and is a much better experience for the students.